Drug utils implementation
